refactor: change PageRepository from class to object
