--- src/data/page/TestPageRepository.ts
@@ -1,13 +1,0 @@
-import { IPageRepository, PageType } from "./IPageRepository";
-
-export class TestPageRepository implements IPageRepository {
-    private _page: PageType = "landing";
-
-    async getCurrentPage(): Promise<PageType> {
-        return this._page;
-    }
-
-    async setCurrentPage(page: PageType): Promise<void> {
-        this._page = page;
-    }
-}
